+++ libc/shared/math.h
@@ -246,6 +246,7 @@
 #include "math/sinpif.h"
 #include "math/sinpif16.h"
 #include "math/sqrt.h"
+#include "math/sqrtbf16.h"
 #include "math/sqrtf.h"
 #include "math/sqrtf128.h"
 #include "math/sqrtf16.h"