+++ libc/src/__support/math/log_range_reduction.h
@@ -15,7 +15,6 @@
 #include "src/__support/uint128.h"
 
 namespace LIBC_NAMESPACE_DECL {
-
 namespace math {
 namespace log_range_reduction_internal {
 
@@ -36,7 +35,7 @@ struct LogRR {
 //   sum: adding -log(r1) - log(r2) - log(r3) - log(r4) to the resulted sum.
 //   return value: the reduced argument v satisfying:
 //                 -0x1.0002143p-29 <= v < 0x1p-29,  and  ulp(v) >= 2^(-125).
-LIBC_INLINE fputil::DyadicFloat<128>
+LIBC_INLINE constexpr fputil::DyadicFloat<128>
 log_range_reduction(double m_x, const LogRR &log_table,
                     fputil::DyadicFloat<128> &sum) {
   using namespace common_constants_internal;
@@ -91,6 +90,7 @@ log_range_reduction(double m_x, const LogRR &log_table,
                               MType({static_cast<uint64_t>(vv4),
                                      static_cast<uint64_t>(vv4 >> 64)}));
 }
+
 } // namespace log_range_reduction_internal
 } // namespace math
 } // namespace LIBC_NAMESPACE_DECL